SLP3
Finally, I make my trek to moving the content form my document to the site. The problems are that I'm not sure how to embed the app into my site, still in the works. For now, the app is linked to different tab. And the last section remains to be finished.